Extract owned_test_projection.rs from turn.rs (parent #680)
Move the Phase-3 (Issue #659 Task 3.3) `owned_test_artifacts_for_verifier`
projection out of `turn.rs` into a focused sibling module so `turn.rs`
keeps shrinking toward responsibility-focused units (parent meta-issue
#680).

One production helper + one private helper extracted as free functions
taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `owned_test_artifacts_for_verifier` (production chokepoint: seeds
  ledger via `artifact_state_projection::task_contract_artifact_states`,
  computes legacy + ledger derivations, emits masked
  `agent.artifact_ledger.divergence_detected` on disagreement, returns
  ledger projection as v0.4.8 production authority)
- `emit_owned_test_artifacts_projection_divergence` (private masked
  observability emit; raw paths never leak, only bounded path-hash
  lists)

Behavior unchanged: same baseline-seeding side effect via
`task_contract_artifact_states` before reading the ledger projection,
same legacy + ledger compare, same masked event shape
(authority="ledger", projection name, bounded path-hash lists).

The Phase-3 signature contract test in
`artifact_ledger_phase3_tests::owned_test_artifacts_for_verifier_signature_unchanged`
was updated to point at the free fn instead of `Agent::...`. The
function-pointer cast still pins the canonical
`fn(&mut Agent, &TaskContract) -> Vec<String>` shape (the signature
itself is unchanged), so a future incompatible refactor still fails
compilation.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `success.rs` (1 site), `verifier_orchestration.rs` (1 site),
`task_contract_recovery.rs` (1 site), and `artifact_ledger_phase3_tests.rs`
(2 sites including the signature pin) to the free-fn form.

LOC delta:
- turn.rs: 1,471 → 1,387 (-84)
- new module: +94

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
