+++ src/agent/loop_run/success.rs
@@ -201,7 +201,8 @@ impl Agent {
         };
         let contract = TaskContract::from_request(&request);
         let test_execution_required = contract.required_behavior.test_execution_required;
-        let owned_test_artifacts = self.owned_test_artifacts_for_verifier(&contract);
+        let owned_test_artifacts =
+            super::owned_test_projection::owned_test_artifacts_for_verifier(self, &contract);
         (owned_test_artifacts, test_execution_required)
     }
