+++ src/agent/loop_run/verifier_orchestration.rs
@@ -2164,7 +2164,8 @@ pub(super) fn task_contract_verifier_test_binding(
     };
     let contract = TaskContract::from_request(&request);
     let test_execution_required = contract.required_behavior.test_execution_required;
-    let owned_test_artifacts = agent.owned_test_artifacts_for_verifier(&contract);
+    let owned_test_artifacts =
+        super::owned_test_projection::owned_test_artifacts_for_verifier(agent, &contract);
     let scope = agent.current_workspace_scope();
     (owned_test_artifacts, test_execution_required, Some(scope))
 }
